moved job scraper into ml.py and cleaned up imports
