+++ app/ml.py
@@ -1,10 +1,13 @@
 """Machine learning functions"""
 from pickle import load
 import requests
+import os
+import datetime
 from bs4 import BeautifulSoup as bs
 from fastapi import APIRouter, HTTPException, Depends
 from pydantic import BaseModel
 from app.state_abbr import us_state_abbrev as abbr
+from dotenv import dotenv_values, load_dotenv
 from pathlib import Path
 import pandas as pd
 import numpy as np
@@ -13,9 +16,8 @@ import asyncio
 from app.db import database, select, select_all
 from typing import List, Optional
 
-
 router = APIRouter()
-
+load_dotenv()
 
 class City(BaseModel):
     city: str = "New York"
@@ -470,3 +472,90 @@ async def get_recommendation_cities(city: City, nearest_string: str):
     )
 
     return recs
+
+# This approach was origianllly developed by https://github.com/israel-dryer/Indeed-Job-Scraper/blob/master/indeed-tutorial.ipynb
+# Also credit https://github.com/jiobu1 for help finding approaches to turning this into an API end point
+# https://www.youtube.com/watch?v=eN_3d4JrL_w
+# https://medium.com/@hannah15198/convert-csv-to-json-with-python-b8899c722f6d
+@router.post('/api/job_opportunities')
+async def job_opportunities(position, city:City):
+    """Returns jobs opportunities from indeed.com
+
+    Fetch first 10 job opportunities
+    - Job title,
+    - Company,
+    - Job location
+    - Post Date,
+    - Extraction Date,
+    - Job Description,
+    - Salary,
+    - Job Url
+    
+    args:
+        - position: desired job
+        - city: target city
+    returns:
+        Dictionary of requested data in a JSON object
+    """
+
+
+    records = []  
+
+    city_name = validate_city(city)
+    location = city_name.city + ' ' + city_name.state
+    url = get_url(position, location) 
+
+    response = requests.get(url)
+    soup = bs(response.text, 'html.parser')
+    cards = soup.find_all('div', 'jobsearch-SerpJobCard')
+
+    for card in cards:
+        record = get_record(card)
+        records.append(record)
+
+    #return total number of jobs by queary and city, state
+    try:
+        total_jobs = soup.find('div', id='searchCountPages').text.strip()
+        total = total_jobs.split()[-2:]
+        jobs = ' '.join(total)
+    except AttributeError:
+        total_jobs = ''
+        jobs = ''
+
+    return {"Search Results":jobs, "Top 10 Listings": records}
+
+def get_record(card):
+    # credit to https://github.com/jiobu1 
+    """Extract job data from a single record"""
+    atag = card.h2.a
+    job_title = atag.get('title')
+    company = card.find('span', 'company').text.strip()
+    job_location = card.find('div', 'recJobLoc').get('data-rc-loc')
+    job_summary = card.find('div', 'summary').text.strip()
+    post_date = card.find('span', 'date').text.strip()
+
+    try:
+        salary = card.find('span', 'salarytext').text.strip()
+    except AttributeError:
+        salary = ''
+
+    extract_date = datetime.datetime.today().strftime('%Y-%m-%d')
+    job_url = 'https://www.indeed.com' + atag.get('href')
+
+    record = {'Job Title': job_title,
+              'Company': company,
+              'Location': job_location,
+              'Date Posted': post_date,
+              'Extract Date': extract_date,
+              'Description': job_summary,
+              'Salary': salary,
+              'Job Url': job_url}
+
+    return record
+
+def get_url(position, location):
+    "Generate a url based on position and location"
+
+    template = "https://www.indeed.com/jobs?q={}&l={}"
+    url = template.format(position, location)
+    return url
